Forward the predicate's status
This is accounted for when calling 'async' and then again when calling 'predicateResult.toBoolean(expectation:)'.

Similar to the functionality in the 'async' equivalent fo the older 'Matcher' API, this should just return 'false' when '.timedOut' occurs, and 'true' when the async'd 'pollBlock' 'isSuccessful'.

However, due to expectation style _also_ being passed into the call to 'async', this was catered for twice, causing a double negative, and essentially creating false positives for all situations.